added cart functionality

+++ browser/js/common/factories/Cart.js
@@ -21,15 +21,26 @@ app.factory('CartFactory', function($http){
 			});
 		},
 
-		removeItem: function (cartId, itemId) {
-			return $http.delete('/api/order/'+cartId, {
-				params: {
-					_id: itemId
-				}
-			})
-			.then(function(response){
-				return response.data;
-			});
+		removeItem: function (cartId, itemId, idx) {
+			if(cartId){
+				return $http.delete('/api/order/'+cartId, {
+					params: {
+						_id: itemId
+					}
+				})
+				.then(function(response){
+					return response.data;
+				});	
+			} else {
+				var removeItem = this.getFromLocalStorage('userCart');
+	            var updatedItem = removeItem.all_items.filter(function(item){
+	                return item.art._id !== itemId;
+	            });
+	            var cart = JSON.parse(localStorage.userCart)
+	            cart.all_items = updatedItem;
+	            localStorage.userCart = JSON.stringify(cart);
+			}
+			
 		},
 
 		submitOrder: function (orderId) {
@@ -41,6 +52,28 @@ app.factory('CartFactory', function($http){
 			.then(function(response){
 				return response.data;
 			});
+		},
+
+		addToLocalStorage: function(cart){
+			if (localStorage.userCart !== undefined){
+				var updatedCart = JSON.parse(localStorage.userCart);
+				updatedCart.all_items.push(cart.all_items[0]);
+				localStorage.userCart = JSON.stringify(updatedCart);
+			} else {
+				localStorage.setItem('userCart', JSON.stringify(cart));
+			}
+			
+		},
+
+		getFromLocalStorage: function(){
+			return JSON.parse(localStorage.getItem('userCart'));
+		},
+
+		createCart: function(cart){
+			console.log('inside createCart')
+			return $http.post('/api/order', {params: cart}).then(function(response){
+				return response.data
+			})
 		}
 
 	}
